Preliminary refactoring
Better acronym for Second Order Perturbation Theory
>> PT -> SOPT

Better function name for Bethe-Hilbert transform
>> semiHilbert -> BetheHilbert

Better names for various control parameters
>> loops -> mloop (*max* number of loops)
>> error -> err (shorter)
>> mixing -> mix (shorter)
>> Resolution -> wres (resolution for the w array)
>> ToLoop -> DoLOOP (clarity)

Added conditional loop-by-loop prints
>> "DMFT-loop #N (max M) has ended", controlled by the quiet flag

Better handling of the strong Mott bias
>> now we avoid to evaluate the BetheHilbert transform if not needed

Much improved handling of the boolean variables activating different kind of spans. Now we have three independent flags:
>> Uline <-> U-driven MIT, given T
>> Tline <-> T-driven MIT, given U
>> UTscan <-> Whole phase diagram, overrides given U and T

Moved almost all the plotting stuff into a "Plotter Module" at the end of the main: we should move it to a proper MATLAB package soon, so to allow calling the plotter functions directly in the REPL.

Gif-related plotting is still in a unlinked script, plot_gifs.m, and will be soon imported in the module, with true gif utilities, not just printing all the frames to png.

--------

WE SHOULD BE READY TO WORK ON THE TODOs