Refactor: optional flags are now strings
Why?

When passing boolean flags, we need to pass a direct true/false value, or initialize somewhere a flag to pass: either way the call is not as clear as I would, especially to use in the REPL. Ideally I'd want something like <function(arg1,arg2... flagname = true), but it seems that MATLAB does not support that.

So here is the solution, just call:

- DMFT_loop(..., 'quiet')
- plot.spectral_frame(..., 'invisible')

NOTE
- In the first case, if the string is not 'quiet', whatever it is (but a string, please) you will not get any error and obtain the default behaviour: all prints go to the REPL. Same if you do not provide any string.
- In the second case instead any string, except 'show', will trigger the default behaviour, i.e. not showing the plot. The same if you do not provide any string. Is it weird? Indeed, but is the most useful behaviour to request a static spectral plot from the REPL: you don't need to remember the exact string and still will see the pictures. It is what it is: spaghetti code. 🥲