Optimize `I_L vs U` calculation: reference status quo
With this state of the code and this input the Luttinger integral vs Hubbard interaction plot saved in showcase/ZvsL_beta1e6_wres2e15.svg is reproduced.

> Quality: excellent
< Performance: terrible

The needed cpu-time, given the extremely high wres=2^15, is unacceptable: ~990s of which ~895s are spent executing the convolutions (built-in conv() function). We may need to optimize this.

- Reducing the calls appear unfeasible: we perform two convolutions for each SOPT call, but the input of the second one is the result of the first one -> no way to vectorize with conv2

- Reducing the self-time per call might be feasible, taking inspiration from the cconv* implementation, that uses built-in FFTs (wrappers of FFTW), but _without_ ensuring the "power of 2" golden rule.

>>> So let's try to build ourselves a pow2conv optimized function!

*) cconv is contained in the signal processing toolbox, and the code is visible; conv is instead a suitable wrapper of the core conv2 matrix function, whose code is obfuscated, so I have no idea how it works.

___

Let's say this opens Issue #1