Optimize `I_L vs U` calculation: fix of math.fconv()
Here we end the story of the mysterious convolution bug.

- Having established that the log plots of the two convolutions are strikingly distinct after the midpoint of the arrays, we go for a careful inspection of that region, that we recall being substantially zero in the bird-eye views found in docs/fast_conv/App.gif
  > we plot just those behaded arrays, with an extremely zoomed y-axis and we see that A. the built-in conv gives a TRUE ZERO (🥺blackmagic!) while B. our fast implementation stays to a totally respectable machine zero. (And we knew it, the norm(conv-fconv) was ridicously small!)

Soooo... are the numerically vanishing negative numbers affecting so much the DMFT self-consistency?! THEY ARE. ✨😶‍🌫️✨

- Let's look downstream in the DMFT_loop: first of all the final result of the nested convolutions (the imag of the 2nd order bubble) is fed to the Kramers-Kronig transform, coded through the hilbert() function embedded in the Signal Processing Toolbox. Will it be so sensible to some vanishing negative numbers?
  > Actually not, the resulting self-energies (docs/fast_conv/sloc_bad.gif) are qualitatively correct, and undistinguishable to the post-fix /sloc_good.gif plots. The error is propagating for sure, but not exploding yet! 🧐

- Then let's continue going downstream and we arrive at the SELF-CONSISTENCY relationship: the Hilbert transform of the Bethe lattice DOS, through the 1/(z-Sigma) Cauchy kernel. This is coded in +phys.BetheHilbert() in the most naive way, directly from the definition of the transform and the analytical expression of the DOS. (maybe at some point we should rewrite it exploiting once more the hilbert() matlab function?)
  > HERE YOU HAVE: docs/fast_conv/gloc_bad.gif is EXTREMELY BAD and totally different from the post-fix /gloc_good.gif file! 😏

So yeah, am I telling you that such a totally negligible 1e-10 difference in the computed convolutions accounts for THIS MESS?! Yes, I am. 😌

The proof is the easiest: add a freaking abs() at the last line of the fconv algorithm (line 60) and... you have it: the magic happens and we retrieve the good old convergence properties and the nicest I_L vs U span (docs/fast_conv/luttinger_good.svg). And with respect to the reference commit 6c38c0d0f6d9f9ecde5a2b81dfabddd76481a7da we go down from ~990s to ~140s, a brillant overall 7x speedup!

We'll provide more insights on the actual speedup of the conv() vs fconv() self-times, but for now we can say that Issue #1 is closed! 💥

Over and out. 🙃