Memento: we really need abs(fconv)
A generic convolution needs to allow negative values. So I wanted to remove the c=abs(c) statement inside the function and craft something more general. I thought that it was a matter of handling the machine zeros. But no, if you completely eliminate the machine zeros you broke again DMFT:

> c(abs(c)<1e-10) = 0)    >>> DOES NOT WORK <<<

So I thought we may just remove the negative machine zeros:

> c(-abs(c)>-1e-10) = 0   >>> AGAIN, NO <<<

Though, this:

> c(-abs(c)>-1e-10) = abs(c(-abs(c)>-1e-10))

actually gives a working result.

But what threshold? 1e-12 does not work, 1e-11 works, but increases significantly the needed number of loops. And we surely expect this to dependent on U and T. A nightmare.

So it appears that *in our specific case* we totally need to make use of the abs() and enforce the convolution to be positive. Sad, but true.