Catchy color prints + beep
'DMFT *not* converged after %d steps\n' will now print in red, without stopping or throwing out a warning. It will beep though. 🙃

With the same strategy also the warnings about convolution error.

NB. it works quite prettily on the MATLAB repl, but remains to see what happens on a terminal and with the standard bash pipelines: the strategy is to send these prints to stderr, so I wonder if we'll find them in the log...