Configure SLURM ARRAY Workflow
The array is constrained to run 1 job at a time: this way is possibile to correctly append the various LOGfiles to a global one.

The SLURM-LOGnames have been modified to comply with the array format.

The $SLURM_ARRAY_TASK_ID environment variable is retrieved *as a double* by MATLAB through the str2double(getenv('SLURM_ARRAY_TASK_ID')) procedure.

Then its value is feed to U [or beta]† as a scaling factor.

    > This way we can span the whole UT surface, and build T-driven [or U-driven]† gifs for each SLURM-dispatched interaction value.

Given the resulting very fine U-grid (U=0.1 in the test input state), I expect to easily retrieve the optimal GIF to display a T-driven Mott transition :)

------------

† We have two new input boolean flags:
    - UARRAY activates SLURM scaling of the U input value
    - TARRAY activates SLURM scaling of the beta input value
      > activating both you can define linear 'oblique' spans!