Start of LATTICES branch
We want to extend the code to include a variety of lattices, so to increase flexibility and explorative power of the package. For particle-hole symmetric dispersions (square, cubic, chain, honeycomb...) it should be quite straightforward and imply just some refactoring of the self-consistency routine, something like creating a +lattice module and moving phys.bethe there, together with the new implementations. Even for the cases where the Gloc expression is not known analytically we could just define the relevant DOS and retrieve numerically the GF, exploting one again the hilbert() function provided by the signal processing toolbox. Though I would avoid it whenever possibile, for efficiency reason: currently the hilbert() calls for the Kramers Kronig transform almost dominate the total CPU time, second only to the fconv calls in the SOPT evaluation.

At some point we would want to deeply(*) modify the code to treat also non p-h symmetric lattices, but let's leave that as a separate, future, task. 

(*) There are many tricks here and there exploiting p-h symmetry currently, with much gain in performance of the code. Also I'm not totally sure the IPT method is even reliable out of p-h symmetry.