FIX: debug on fermi function was ill defined
if ( fh ~= f ) was a very unfortunate way to evaluate the if condition for an *array* of logicals. MATLAB evaluates to false an array that has *at least a zero value*. Hence:

if fh == f
   < stuff >
else
   < error >
end

would have worked: if at least one element of fh and f differ, the if evaluates to false and we get an error. Exchanging the if and else fields instead gives unintended behavior.

>>> We'd better avoid using such a slippery vectorial behavior for logical conditions, make use of a much more clearly any() / all() approach. In this case:

if any( fh ~= f )
   < error >
else
   < stuff >
end