First try at square lattice
Works, but it poses loads of problems:

- The Symbolic Math Toolbox function ellipticK() is soooooo muuuuuuch sloooooooooooooooow, its computing time dominates everything. I guess I should have forseen this: a symbolic implementation to process very dense numerical samples of complex numbers it's such a bad idea. Overall a bethe lattice loop takes ~300ms, while a square lattice loop around 24s, that is a x80 slowdown. Damn it.

- Anyway I'd prefer to avoid forcing yet another toolbox dependency.

HENCE

Could we try the good old fft-powered numerical way?

1. Generate the DOS by whatever, numerical, method (even elliptic integrals, but numerical, probably the standard ellipke() MATLAB function). If things go really bad fuck it, let's generate it one time and store a binary file in the repository.

2. Go by definition of self-consistency: Gloc(z) = H[DOS](z-Sloc(z)), an Hilbert transform to which we feed the DOS (generated or tabulated, whatever). But how the hell we evaluate a transform at given z-Sigma(z)? THIS COULD BE AN INEVITABLE PROBLEM.

NB1) Frustration apart, we'd try to avoid tabulating the DOSes, we want flexibility with the energy sampling and interpolating / downsampling could be cumbersome.

NB2) If we succeed generating the DOS we should consider avoiding regenerating at each loop. Instead of passing around a string identifying the lattice we should then consider passing directly the generated DOS.

NB3) This has to perform better than the symbolic method in gftools, BUT it could still be significantly slower / less accurate than the current implementation, for the Bethe lattice. Maybe let's make the DOS argument optional and fall back to the bethe lattice by default. This way the blazing-fast, reliable, infinite-dimensions implementation would be preserved.