Checkpoint: Luttinger integral works for 1d-chain too :fire:
Looking at Z(U) and at the evolution of the spectral functions (both the DOS and the Self-Energy) the zero-T MIT seems perfectly† captured. The Luttinger integral displays a fairly good first order step at the transition and the integer value in the Mott phase is the same of the Bethe lattice.

† Actually Uc seems a little suspicious, maybe we have some problems with the scaling of U in the plots.