IMPORTANT FIX: U/D scaling in *calculations*
The generic equation for the Weiss field is:

g0(z) = ( z + mu - hybr )^-1
      = ( z + mu - t^2 ∑[G(k,z)] )^-1

In our case we have

> mu=0 (half-filling + ph-sym)
> ∑[G(k,z)] = Gloc(z), where Gloc(z) is computed by phys.gloc(z,D,dos)

For the Bethe lattice t = D/2, so that the *HARDCODED MAGIC NUMBER* 🙈 t^2 = 0.25 was somewhat appropriate†, but for sure for a generic lattice this won't work. Indeed we found very strange behavior varying D on the square and cubic lattices, namely different A(w) shapes for a fixed U,D pair, different Uc/D for different D values and all this kind of crazy stuff.

So we fix this by

- defining a proper hybr(D,DOS,GLOC) function, which just fills in the correct t(D) relationship for all the lattices
  > ugly, but for now I don't know how to improve it... maybe after all we really need a +lattice module, I'll think about it.
- building the Weiss field as g0 = ( z - hybr(...))^-1

Preliminary testing has confirmed that

- Uc/D is now fixed 🥲
- Z vs U/D plots are now independent on D 😮‍💨
- Same for spectral plots and gifs (except the bandwidth, of course) 🙃

---------

†Actually, even for the Bethe lattice, changing D would break something, since (D/2) would still differ from 1/4. We should go fix this in the main to, in order to have a proper 'Bethe-only' release.