FIX: Luttinger integral for odd wres (w=0 divergence)
The original definition, as given by Logan, Tucker and Galpin, prescribes
to integrate over the negative semiaxis:

    I = \frac{2}{\pi} \Im\int_{-\infty}^{0} dw G_loc(w) d\Sigma(w) / dw

Nevertheless we find most useful to exploit particle-hole symmetry and
thus integrate over the whole real-axis. Care has to be taken whenever
the frequency domain contains the origin, for therein resides a nasty
pole, arising from the derivative of the self-energy. Thus we divide
the integrand in (w < 0) and (w > 0) parts, and integrate the union.
Failing to exclude the (w = 0) point, if exist, leads to a diverging
uncontrolled result.

-------

Now the odd wres case should be totally safe.