Add density + double occupancy + local entropy
Tested on Bethe and square lattices. It works, but requires VERY high cutoff
on imaginary axis to give accurate results, so much to render performance
critical the analytic continuation performed in phys.matsubara. We are talking
about vcut≈50 wrt the expected vcut≈10. This means dealing with ≈80k matsubara
frequencies at zero temperature, which just renders impractical a direct 
summation-on-all-the-real-frequencies algorithm to evaluate the G(iv) and S(iv)
at each iv value. We really need to speed it up somehow to avoid it becoming
a serious bottleneck. 


Recalling that vectorization seems of no help whatsoever (cfr. bd6726b), 
we might want to consider the use of adaptive quadrature integration,
which would amount to define a proper interpolation scheme in real frequency,
for the integrand function. Tedious, but manageable I believe.