Checkpoint: nice local entropy line benchmark [against literature]
The current code reproduces nicely fig. 1 in Mod. Phys. Lett. B 27:05,
aside the well known overestimation of Uc2 (the paper reports ED results)
we have the same qualitative shape in the FL phase and a fair enough value
in the Mott phase. It appears to saturate to 1.2*{theoretical value} as a
consequence of a too large double occupancy (it never reaches zero, due to
the well known "tail-problem" in the matsubara summation. Again, we need to
either substantially speed-up phys.matsubara, so to allow longer tails, or
to find some sort of tail correction to the migdal-galitskij summation.