Adaptive quadrature is NOT the solution
Speeding up phys.matsubara is rapidly becoming a nightmare.
Using adaptive quad integration, even in the most trivial case (U=0, high
temperature, everything smooth, everything easy) is SLOW, terribly slow 🐢

We are talking 98-99% of the total elapsed time, spent into integral(),
half of which interpolating the spectral function A(ω). And we are even
"interpolating" with the 'nearest' scheme, which is bad and of course the
fastest possible. We are also requiring a ridiculous 1e-3 tolerance on the
integral... so I guess that no, quad integration is indeed not the solution.

☠️