Second attempt at vectorization oh phys.matsubara... no way!
We have already tried a full vectorization of the algorithm (matrix products)
> cfr. bd6726b58db69178d9d53fb47cf54b7b4fea333f

It was significantly slower then the naive loop of vector operations.

Now we try a different (I'd say more elegant) vectorization scheme, inspired
by previous commit (987ed33), which amounts to define a function handle for
generic complex argument z, dFiv(z), and pass to it the matsubara frequencies
all together when summing: sum(dFiv(1i*v)). Currently this is the best timing
we could achieve, but still slower (x5) than the naive loop over vectors.

The first, direct, implementation could really be the fastest possible...
which is indeed a problem for our purposes. We might really need to find
a suitable tail correction for the matsubara sums, so to be able to keep
vcut at least an order of magnitude slower than the current one (which would
mean of the same of order of wcut, which sounds definitely good...).

Let's see...