Dynamic threshold for Dirac velocity estimation [to be tuned]
- The very small (0.01, the same as) w_th we used to fit
- Increasing w_th from 0.01 to 0.1 fixes the low-coupling interacting
  points but totally breaks the strong coupling ones
- Looking at the debug plot we see that the problem is the very sensible
  position of the van Hove singularities in the interacting DOS
> So we need indeed a "dynamic" threshold to be determined wrt those van
  Hove peaks (wmax).

Trying w_th = wmax, wmax/2, wmax/3, wmax/10 gives vey different profiles
ranging from nonmonotonic ones, akin to what happens with the zetaweight
to totally broken mid-coupling cases, to something similar to the ref in
literature (Jafari, 2009), but not quite there quantitatively. I suppose
the tuning of <a> in wmax/<a> to be quite subtle.

------

NB. Is quite possible that the problem we face with the zeta weight is
    the very same in this scenario. Probably worth some test.