FIX: Z weight for the honeycomb lattice
Following the hints emerged within commits [1] and [2] we change here the
threshold for the fitting window within phys.zetaweight and actually recover
a qualitative sane behavior for the U-driven Mott transition. This opens
many question about the reliability of the fitting procedure (at least
without rigorous checks on the threshold). We should really return tuning
the other lattices and maybe compare with the matsubara evaluation of Z,
so moving the testing to the matsubara branch.

[1]: 1ab848f24174484d473d03d69e00cc21d4e0e563
[2]: f8a5b53e598cdbafefeb8430e252982763415831