Retcon commit 81bd520ed8f0b59f529ce68678c8889d29834cc3
Since that fix we had a coherent D≠1 behavior, but very suspicious Uc
values for well-known lattices: mainly square lattice resulting in half
the DMFT/ED Uc (and IPT notoriously overestimates critical interactions)
and honeycomb lattice being at 1/3 of the IPT value reported by Jafari.

More fog has arisen while playing with Economou's 2d toy-model, so we
ended up losing trust in the lattice-generic implementation.

After some careful testing and literature browsing we determine that the
right expression for the hybridization function is *not* t^2 ∑[G(k,z)],
but (D/2)^2 ∑[G(k,z)], which matches with the former in the Bethe lattice
case, thus explaining why Rozenberg reports it in that form in his notes.

Hence here we:

- add a digital note to embedded Rozenberg pdf to highlight the mistake

- fix the Weiss field expression in dmft_loop() function

- move the t(D) table from the private hybr() function to phys.hopping()
  with the idea of exploiting it to give the user the option to plot in
  U/t units, instead of the default (for sure appropriate) U/D.

I feel like we could restart trusting the lattice extension of the code
and approach a merge-to-master phase.