Import ADAPTIVE MIXING scheme from SciFortran
For now an almost literal translation of
> https://github.com/QcmPlab/SciFortran/tree/master/src/SF_OPTIMIZE/adaptive_mix.f90

Hence it runs *very* slow (it dominates the runtime of dmft_loop, taking slightly
more time than the SOPT. This is due to the explicit for loop on the real freqs,
which needs then to be vectorized for serious usage (and careful testing already).

It definitely improves the U=2.5 benchmark, but still we cannot really take over the
U=3.3 barrier. Nevertheless more testing is needed and that requires vectorization.

-------------------------------------------------------------------------------------

Why not Broyden?

Well, standard Broyden mixing requires storing a NxN Jacobian matrix, where N is the
number of real frequencies in our case. Trying allocating that results in a disaster:

   Requested 32768x32768 (16.0GB) array exceeds
   maximum array size preference. Creation of arrays
   greater than this limit may take a long time and
   cause MATLAB to become unresponsive.

Srivastava 'mxN improvement' would then be necessary, but it carries along important
drawbacks, leading to further modifications (Vanderbilt-Louie-Johnson scheme) that are
probably too much for a first try. We might return to it later...

More info:
> https://journals.aps.org/prb/abstract/10.1103/PhysRevB.38.12807 (Johnson)
> https://journals.aps.org/prb/abstract/10.1103/PhysRevB.80.125125 (Žitko)