SOFT-FIX dmft_loop (formally incorrect call to adaptive mixing)
Adaptive mixing algorithm is devised for real input arrays x and Fx, but
we were passing to it Weiss fields (and their differences). Matlab gives
undefined behavior when comparing complex numbers to zero (I suspect it
actually compares the real parts, but docs say > and < are not supported
for complex data)
> https://it.mathworks.com/help/stateflow/ug/operations-on-complex-data.html

So now we explicitly mix the imaginary part of the Weiss field and take
a Kramers Kronig transform to retrieve the whole complex mixed function.

The KK transform can be actually avoided since all we actually need for
the SOPT is the spectral function of the Weiss field... I'll refactor :)

NB. Convergence seems to be further improved, but Mott still a problem.