Checkpoint: move restart, mixing and error check from g0 to A0
This does not work, since working (especially mixing) only with the imag
part of the Weiss field makes manipulations causality-unsafe. The best
thing to do is probably extend the adaptive mixing to take complex input.

The current complex implementation in SciFortran is broken, as far as I
can tell, so I have to think about a fix.