Revert to simple mixing
Whenever there is significant redistribution of spectral weight, as it
happens near the Mott transition, the adaptive mixing scheme may violate
the normalization of the spectral function. This leads to catastrophic
behavior in convergence. We keep it in the codebase, for it has proven
possibly useful in the Fermi liquid, but for now let's keep it uncalled.